pyproject.toml
[tool.poetry]
name = "yao"
version = "0.2.0"
description = "An implementation coming right from metaverse of the most amazing dog"
authors = ["THhéophane Vié <theophane.vie@petit-nuage.org>"]

[tool.poetry.dependencies]
python = "^3.11"
cowsay = "^5.0"
youtube-dl = {git = "https://github.com/ytdl-org/youtube-dl.git"}
discord-py = {extras = ["voice"], version = "^2.3.2"}

[tool.poetry.dev-dependencies]
ruff = "^0.3.0"

[tool.ruff]
ignore = ["F503", "E203"]
extend-select = [
  "E", # pycodestyle
  "W", # pycodestyle
  "I", # isort
  "C", # complex-structure
]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
